requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.2.0
Django==4.1.6
django-environ==0.9.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
drf-yasg==1.21.4
Faker==16.6.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pylama==8.4.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
